Module ncollide_entities::partitioning

Spatial partitioning tools.

Structs

BVT
A Boundig Volume Tree.

BoundingVolumeInterferencesCollector
Bounding Volume Tree visitor collecting interferences with a given bounding volume.

DBVT
A Dynamic Bounding Volume Tree.

DBVTLeaf
Leaf of a Dynamic Bounding Volume Tree.

Enums

BVTNode
A node of the bounding volume tree.

BinaryPartition
Result of a binary partition.

Traits

BVTCostFn
Trait implemented by cost functions used by the best-first search on a BVT.

BVTTVisitor
Visitor for the Bounding Volume Traversal Tree.

BVTVisitor
Visitor of Bounding Volume Trees.

Functions

median_partitioner
Construction function for a kdree to be used with BVT::new_with_partitioner.

median_partitioner_with_centers
Construction function for a kdree to be used with BVT::new_with_partitioner.