Struct ncollide_entities::shape::Reflection

pub struct Reflection<'a, G: ?Sized + 'a> {
    // some fields omitted
}

SupportMap representation of the reflection of a shape.

A reflection is obtained with the central symmetry with regard to the origin.

Methods

impl<'a, G: ?Sized> Reflection<'a, G>

fn new(shape: &'a G) -> Reflection<'a, G>

Build the reflection of a shape. Since the representation is implicit, the reflection computation is done in constant time.

fn shape(&self) -> &'a G

The reflected shape.