Module ncollide_entities::support_map

Definition of support functions.

Traits

PreferedSamplingDirections
Trait of shapes having prefered sampling directions for the Minkowski sampling algorithm.

SupportMap
Traits of convex shapes representable by a support mapping function.

Functions

cso_support_point
Computes the support point of the CSO g1 - g2 on a given direction.

point_cloud_support_point
Computes the support point of a cloud of points.