Trait ncollide_entities::support_map::SupportMap

pub trait SupportMap<P: Point, M> {
    fn support_point(&self, transform: &M, dir: &P::Vect) -> P;
}

Traits of convex shapes representable by a support mapping function.

Parameters:

V - type of the support mapping direction argument and of the returned point.

Required Methods

fn support_point(&self, transform: &M, dir: &P::Vect) -> P

Evaluates the support function of the object. A support function is a function associating a vector to the shape point which maximizes their dot product. This does not include the margin of the object. Margins are shape-dependent. Use support_point to sample the complete shape.

Arguments:

dir - the input of the support function. It is not required for it to be normalized.