Module ncollide_entities::inspection

Traits and methods to inspect and retrieve the capabilities and representations of shapes at runtime.

Reexports

pub use self::repr::{ReprDesc, Repr};

Functions

composite_shape_repr_id
Gets the id associated with the CompositeShape trait.

maybe_as_composite_shape
Converts a shape to a composite shape if possible.

maybe_as_support_map
Converts a shape to a support map if possible.

maybe_repr_desc_as_composite_shape
Converts a shape to a composite shape if possible.

maybe_repr_desc_as_support_map
Converts a shape descriptor to a support map if possible.

support_map_repr_id
Gets the id associated with the SupportMap trait.

Type Definitions

Repr2
A 2d dynamic representation object.

Repr3
A 3d dynamic representation object.