Struct ncollide_entities::partitioning::BVT

pub struct BVT<B, BV> {
    // some fields omitted
}

A Boundig Volume Tree.

Methods

impl<B, BV> BVT<B, BV>

fn new_with_partitioner<F: FnMut(usize, Vec<(B, BV)>) -> (BV, BinaryPartition<B, BV>)>(leaves: Vec<(B, BV)>, partitioner: &mut F) -> BVT<B, BV>

Builds a bounding volume tree using an user-defined construction function.

fn visit<Vis: BVTVisitor<B, BV>>(&self, visitor: &mut Vis)

Visit this tree using… a visitor!

fn visit_bvtt<Vis: BVTTVisitor<B, BV>>(&self, other: &BVT<B, BV>, visitor: &mut Vis)

Visits the bounding volume traversal tree implicitely formed with other.

fn best_first_search<'a, N, BFS, R>(&'a self, algorithm: &mut BFS) -> Option<(&'a B, R)> where N: Scalar, BFS: BVTCostFn<N, B, BV, R>

Performs a best-fist-search on the tree.

Returns the content of the best leaf nound, and a result of user-defined type.

fn root_bounding_volume<'r>(&'r self) -> Option<&'r BV>

Reference to the bounding volume of the tree root.

fn depth(&self) -> usize

Computes the depth of this tree.

impl<N, V, B, BV> BVT<B, BV> where N: Scalar, V: Vect<N>, BV: Translation<V> + BoundingVolume<N> + Clone

fn new_balanced(leaves: Vec<(B, BV)>) -> BVT<B, BV>

Creates a balanced BVT.